requirements.txt
# LLM & Embeddings
google-genai>=1.0.0

# Vector DB
faiss-cpu>=1.8
numpy>=1.26

# Document Parsing
pypdf>=4.0
langchain-text-splitters
python-docx>=1.1
unstructured>=0.13
tiktoken>=0.7

# Backend
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
pydantic-settings>=2.0
slowapi>=0.1
httpx>=0.27
aiosqlite>=0.20

# Validation
jsonschema>=4.0

# Utils
python-dotenv>=1.0
